+++ apollos-church-api/src/data/wcc-blog/index.js
@@ -1,1 +1,5 @@
-export * as WCCBlog from './blog-items';
+import resolver from './resolver';
+import dataSource from './data-source';
+import schema from './schema';
+
+export { resolver, dataSource, schema };
