+++ apollos-church-api/src/data/wcc-blog/schema.js
@@ -1,0 +1,33 @@
+import gql from 'graphql-tag';
+
+const schema =  gql`
+  type WCCBlog implements ContentItem & Node {
+    id: ID!
+    title(hyphenated: Boolean): String
+    coverImage: ImageMedia
+    htmlContent: String
+    summary: String
+    childContentItemsConnection(
+      first: Int
+      after: String
+    ): ContentItemsConnection
+    siblingContentItemsConnection(
+      first: Int
+      after: String
+    ): ContentItemsConnection
+    images: [ImageMedia]
+    videos: [VideoMedia]
+    audios: [AudioMedia]
+    parentChannel: ContentChannel
+    theme: Theme
+  }
+
+  extend type Query {
+    blogs(
+      first: Int
+      after: String
+    ): ContentItemsConnection
+  }
+`;
+
+export default schema;
