+++ apollos-church-api/src/data/wcc-media/index.js
@@ -1,1 +1,5 @@
-export * as WCCMessage from './messages';
+import resolver from './resolver';
+import dataSource from './data-source';
+import schema from './schema';
+
+export { resolver, dataSource, schema };
